Fixed Object Class counts and standardized queryset filters
