+++ usaspending_api/disaster/v2/views/agency/count.py
@@ -2,15 +2,15 @@ from django.db.models import OuterRef, Q, Exists, Count
 from rest_framework.request import Request
 from rest_framework.response import Response
 
-from usaspending_api.common.cache_decorator import cache_response
-from usaspending_api.disaster.v2.views.count_base import CountBase
 from usaspending_api.awards.models import FinancialAccountsByAwards
+from usaspending_api.common.cache_decorator import cache_response
+from usaspending_api.disaster.v2.views.disaster_base import AwardTypeMixin
+from usaspending_api.disaster.v2.views.disaster_base import DisasterBase
 from usaspending_api.financial_activities.models import FinancialAccountsByProgramActivityObjectClass
 from usaspending_api.references.models import ToptierAgency
-from usaspending_api.disaster.v2.views.disaster_base import AwardTypeMixin
 
 
-class AgencyCountViewSet(AwardTypeMixin, CountBase):
+class AgencyCountViewSet(AwardTypeMixin, DisasterBase):
     """
     Obtain the count of Agencies related to supplied DEFC filter.
     """
@@ -19,25 +19,22 @@ class AgencyCountViewSet(AwardTypeMixin, CountBase):
 
     @cache_response()
     def post(self, request: Request) -> Response:
+        filters = [
+            self.all_closed_defc_submissions,
+            self.has_award_of_provided_type,
+            self.is_in_provided_def_codes,
+            self.is_non_zero_award_spending,
+        ]
+
         if self.award_type_codes:
-            filters = [
-                self.is_in_provided_def_codes(),
-                self.all_closed_defc_submissions,
-                self.is_non_zero_award_spending(),
-            ]
             count = (
                 FinancialAccountsByAwards.objects.filter(*filters)
                 .values("award_id")
-                .filter(award__type__in=self.filters.get("award_type_codes"))
                 .aggregate(count=Count("award__funding_agency__toptier_agency", distinct=True))["count"]
             )
+
         else:
-            filters = [
-                Q(treasury_account__funding_toptier_agency=OuterRef("pk")),
-                self.is_in_provided_def_codes(),
-                self.all_closed_defc_submissions,
-                self.is_non_zero_total_spending(),
-            ]
+            filters.append(Q(treasury_account__funding_toptier_agency=OuterRef("pk")))
             count = (
                 ToptierAgency.objects.annotate(
                     include=Exists(FinancialAccountsByProgramActivityObjectClass.objects.filter(*filters).values("pk"))
@@ -46,4 +43,5 @@ class AgencyCountViewSet(AwardTypeMixin, CountBase):
                 .values("pk")
                 .count()
             )
+
         return Response({"count": count})
