+++ usaspending_api/disaster/v2/views/award/count.py
@@ -2,13 +2,13 @@ from django.db.models import Count
 from rest_framework.request import Request
 from rest_framework.response import Response
 
-from usaspending_api.common.cache_decorator import cache_response
-from usaspending_api.disaster.v2.views.count_base import CountBase
 from usaspending_api.awards.models import FinancialAccountsByAwards
+from usaspending_api.common.cache_decorator import cache_response
+from usaspending_api.disaster.v2.views.disaster_base import DisasterBase
 from usaspending_api.disaster.v2.views.disaster_base import FabaOutlayMixin, AwardTypeMixin
 
 
-class AwardCountViewSet(CountBase, FabaOutlayMixin, AwardTypeMixin):
+class AwardCountViewSet(DisasterBase, FabaOutlayMixin, AwardTypeMixin):
     """
     Obtain the count of Awards related to supplied DEFC filter.
     """
@@ -21,9 +21,9 @@ class AwardCountViewSet(CountBase, FabaOutlayMixin, AwardTypeMixin):
     def post(self, request: Request) -> Response:
 
         filters = [
-            self.is_in_provided_def_codes(),
+            self.is_in_provided_def_codes,
             self.all_closed_defc_submissions,
-            self.is_non_zero_award_spending(),
+            self.is_non_zero_award_spending,
         ]
         count = FinancialAccountsByAwards.objects.filter(*filters)
         if self.award_type_codes:
