+++ usaspending_api/disaster/v2/views/federal_account/count.py
@@ -4,11 +4,11 @@ from rest_framework.response import Response
 
 from usaspending_api.accounts.models import FederalAccount
 from usaspending_api.common.cache_decorator import cache_response
-from usaspending_api.disaster.v2.views.count_base import CountBase
+from usaspending_api.disaster.v2.views.disaster_base import DisasterBase
 from usaspending_api.financial_activities.models import FinancialAccountsByProgramActivityObjectClass
 
 
-class FederalAccountCountViewSet(CountBase):
+class FederalAccountCountViewSet(DisasterBase):
     """
     Obtain the count of Federal Accounts related to supplied DEFC filter.
     """
@@ -19,9 +19,9 @@ class FederalAccountCountViewSet(CountBase):
     def post(self, request: Request) -> Response:
         filters = [
             Q(treasury_account__federal_account_id=OuterRef("pk")),
-            self.is_in_provided_def_codes(),
+            self.is_in_provided_def_codes,
             self.all_closed_defc_submissions,
-            self.is_non_zero_total_spending(),
+            self.is_non_zero_total_spending,
         ]
         count = (
             FederalAccount.objects.annotate(
