+++ usaspending_api/disaster/v2/views/object_class/spending.py
@@ -49,15 +49,10 @@ class ObjectClassSpendingViewSet(PaginationMixin, SpendingMixin, DisasterBase):
     @property
     def total_queryset(self):
         filters = [
-            Q(
-                Q(obligations_incurred_by_program_object_class_cpe__gt=0)
-                | Q(obligations_incurred_by_program_object_class_cpe__lt=0)
-                | Q(gross_outlay_amount_by_program_object_class_cpe__gt=0)
-                | Q(gross_outlay_amount_by_program_object_class_cpe__lt=0)
-            ),
-            Q(disaster_emergency_fund__in=self.def_codes),
-            Q(object_class__isnull=False),
+            self.is_in_provided_def_codes,
+            self.is_non_zero_total_spending,
             self.all_closed_defc_submissions,
+            Q(object_class__isnull=False),
         ]
 
         annotations = {
@@ -101,7 +96,7 @@ class ObjectClassSpendingViewSet(PaginationMixin, SpendingMixin, DisasterBase):
     @property
     def award_queryset(self):
         filters = [
-            Q(disaster_emergency_fund__in=self.def_codes),
+            self.is_in_provided_def_codes,
             Q(object_class__isnull=False),
             Q(award__isnull=False),
             self.all_closed_defc_submissions,
